Document the bootstrap process.
This change adds a bunch of docs describing the bootstrap process.  It also
renames a couple things in the source to better correspond with the terminology
used in the docs.  On top of that it makes the standalone build to copy minibp
from the .bootstrap/bin directory to the bin directory to make it easier to
find.

Change-Id: Ibae3e607b59e2caf7d764ec3153155fde208ec1e