gotestrunner: Make GOROOT absolute before chdir
If GOROOT is a relative path (like it is in Android), then changing
directories to run the test will cause GOROOT to become invalid. So make
it absolute if we're going to change to a different directory.

This allows tests to use the go tools without searching for a valid go
installation.

Change-Id: Ifab0a8533d236054ccf363dfb68b12e0bf66f6f8