Preemptively rename File to ParseTree to clarify its upcoming meaning.
A ParseTree represents a tree that was parsed from text, and also
includes information about where its nodes came from. Currently the
source information is attached to the nodes themselves, but that's
changing in a subsequent patch, at which point the ParseTree will
directly contain the source information. At that point, it will be
possible for a programmatic builder class to generate a syntax tree
without having to think about or pass in empty values for text
positions.

Really the reason that this patch is separate is to make the code review
more convenient.

gorename -from 'github.com/google/blueprint/parser.File.Name' -to 'FileName'
gorename -from 'github.com/google/blueprint/parser.File' -to 'ParseTree'

Bug: b/34671474
Test: make blueprint_tools

Change-Id: I7d591ac123ccd1697a3d739308596655afc39197